refacto examplo to jlianes package
