+++ app/google-services.json
@@ -41,6 +41,43 @@
           ]
         }
       }
+    },
+    {
+      "client_info": {
+        "mobilesdk_app_id": "1:793109226456:android:ab2c0a7261cb2d62b9c8ad",
+        "android_client_info": {
+          "package_name": "com.jlianes.birthdaynotifier"
+        }
+      },
+      "oauth_client": [
+        {
+          "client_id": "793109226456-iv5v8ifiobvjr25gabk3r5o6bcoqii08.apps.googleusercontent.com",
+          "client_type": 1,
+          "android_info": {
+            "package_name": "com.jlianes.birthdaynotifier",
+            "certificate_hash": "df95a6fd69c6ea66742fd2e4266db0eab28341c0"
+          }
+        },
+        {
+          "client_id": "793109226456-n35uqu21dd5s3i3ggn57807tvlkrgio9.apps.googleusercontent.com",
+          "client_type": 3
+        }
+      ],
+      "api_key": [
+        {
+          "current_key": "AIzaSyCcMGFQT7hBJbvqokzvYInqBxeoRdH0XdY"
+        }
+      ],
+      "services": {
+        "appinvite_service": {
+          "other_platform_oauth_client": [
+            {
+              "client_id": "793109226456-n35uqu21dd5s3i3ggn57807tvlkrgio9.apps.googleusercontent.com",
+              "client_type": 3
+            }
+          ]
+        }
+      }
     }
   ],
   "configuration_version": "1"
