--- app/src/main/java/com/example/birthdaynotifier/framework/receiver/AlarmScheduler.kt
@@ -1,42 +1,0 @@
-package com.example.birthdaynotifier.framework.receiver
-
-import android.app.AlarmManager
-import android.app.PendingIntent
-import android.content.Context
-import android.content.Intent
-import java.util.Calendar
-
-/** Helper object to schedule the daily birthday check alarm. */
-object AlarmScheduler {
-    /** Schedule the alarm based on saved time preferences. */
-    fun schedule(context: Context) {
-        val prefs = context.getSharedPreferences("settings", Context.MODE_PRIVATE)
-        val hour = prefs.getInt("hour", 9)
-        val minute = prefs.getInt("minute", 0)
-
-        val alarmManager = context.getSystemService(Context.ALARM_SERVICE) as AlarmManager
-        val intent = Intent(context, BirthdayReceiver::class.java)
-        val pendingIntent = PendingIntent.getBroadcast(
-            context,
-            0,
-            intent,
-            PendingIntent.FLAG_IMMUTABLE
-        )
-
-        val calendar = Calendar.getInstance().apply {
-            set(Calendar.HOUR_OF_DAY, hour)
-            set(Calendar.MINUTE, minute)
-            set(Calendar.SECOND, 0)
-            if (before(Calendar.getInstance())) {
-                add(Calendar.DAY_OF_MONTH, 1)
-            }
-        }
-
-        alarmManager.setInexactRepeating(
-            AlarmManager.RTC_WAKEUP,
-            calendar.timeInMillis,
-            AlarmManager.INTERVAL_DAY,
-            pendingIntent
-        )
-    }
-}
